all: Move assets packages around
lib/api/assets => lib/assets
lib/auto => lib/api/auto
