+++ lib/api/auto/doc.go
@@ -4,7 +4,7 @@
 // License, v. 2.0. If a copy of the MPL was not distributed with this file,
 // You can obtain one at https://mozilla.org/MPL/2.0/.
 
-//go:generate go run ../../script/genassets.go -o gui.files.go ../../gui
+//go:generate go run ../../../script/genassets.go -o gui.files.go ../../../gui
 
 // Package auto contains auto generated files for web assets.
 package auto
