+++ lib/assets/assets.go
