+++ lib/assets/assets_test.go
